+++ index.js
@@ -113,6 +113,14 @@ module.exports = class ScreepsModules {
           return reject(err)
         }
 
+        if (response.statusCode === 401) {
+          this.options.token = ''
+        }
+
+        if (response.headers['x-token']) {
+          this.options.token = response.headers['x-token']
+        }
+
         if (body.hasOwnProperty('error')) {
           return reject(body.error)
         }