+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 ### Added
+- Invalidate token on 401 responses
+- Always set token found response headers
 - Improved error handling
 
 ## [1.0.2] - 2016-11-13